% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{choose_time}
\alias{choose_time}
\title{Pick a Convenient Time}
\usage{
choose_time(book_name, facilitator_id, required_choosers = 5)
}
\arguments{
\item{book_name}{The abbreviation for an approved book, such as rpkgs or
mshiny.}

\item{facilitator_id}{The Slack ID of the facilitator for this group.
Available via the user's "full profile" in Slack.}

\item{required_choosers}{The minimum number of sign-ups necessary for the
club to launch. Think really hard before changing this.}
}
\value{
The best times, as a tibble.
}
\description{
There's a lot of work to do on this still.
}
