% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_global.R
\name{.app_global}
\alias{.app_global}
\title{Stuff to run at startup}
\usage{
.app_global()
}
\description{
I'll have to see what should universally run at startup (eg, definitions of
the possible slots) vs what should be checked during use (eg, we should load
the "used" slots from time to time, although we can probably do that in
global with a reactivePoll, so... yeah, this will get more stuff).
}
\keyword{internal}
