% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_global.R
\name{.authorize_google_services}
\alias{.authorize_google_services}
\title{Authorize Google for Sheets and Drive}
\usage{
.authorize_google_services()
}
\value{
NULL (invisibly).
}
\description{
Authorize Google for Sheets and Drive
}
\keyword{internal}
