% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{.book_observer}
\alias{.book_observer}
\title{Load Books and Set Dropdown Values}
\usage{
.book_observer(
  id = "book_name",
  approved_books,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{The identifier of this input.}

\item{approved_books}{A data.frame of approved books with column `book_name`.}

\item{session}{The shiny session. I don't know of a case where we shouldn't
use the default.}
}
\value{
An observer that sets up the book drop-down menu.
}
\description{
Load Books and Set Dropdown Values
}
\keyword{internal}
