% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{.book_server}
\alias{.book_server}
\title{Load Books and Set Dropdown Values}
\usage{
.book_server(id = "book_name", session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The shiny session. I don't know of a case where we shouldn't
use the default.}
}
\value{
An observer that sets up the book drop-down menu.
}
\description{
Load Books and Set Dropdown Values
}
\keyword{internal}
