% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{.book_ui}
\alias{.book_ui}
\title{UI to Choose a Book}
\usage{
.book_ui(id = "book_name")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
A shiny module to select a book.
}
\description{
UI to Choose a Book
}
\keyword{internal}
