% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.calendar_selector}
\alias{.calendar_selector}
\title{Server-side Calendar Selector}
\usage{
.calendar_selector(user_id, user_timezone, signups)
}
\arguments{
\item{user_id}{Slack's unique id for this user.}

\item{user_timezone}{The timezone currently chosen by the user (as a
character scalar).}

\item{signups}{A dataframe of signup information, converted to UTC.}
}
\value{
A [rhandsontable::renderRHandsontable()].
}
\description{
Server-side Calendar Selector
}
\keyword{internal}
