% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.calendar_ui}
\alias{.calendar_ui}
\title{UI to Choose Times}
\usage{
.calendar_ui(id = "calendar")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
A [rhandsontable::rHandsontableOutput()] UI.
}
\description{
UI to Choose Times
}
\keyword{internal}
