% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{.check_books}
\alias{.check_books}
\title{Check Club Sheet Modified Time}
\usage{
.check_books()
}
\value{
A string representing when the sheet was modified.
}
\description{
Check Club Sheet Modified Time
}
\keyword{internal}
