% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{.check_club_sheet}
\alias{.check_club_sheet}
\title{Check Club Sheet Modified Time}
\usage{
.check_club_sheet()
}
\value{
A list with components of when the sheet was modified and the size of
  the sheet. It seems that this combo triggers updates more accurately than
  just the modifiedTime alone.
}
\description{
Check Club Sheet Modified Time
}
\keyword{internal}
