% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.generate_renderer_js}
\alias{.generate_renderer_js}
\title{Generate JavaScript for formatting the table}
\usage{
.generate_renderer_js(density_matrix)
}
\arguments{
\item{density_matrix}{The signup information (in this user's timezone) as a
matrix.}
}
\value{
JavaScript code as a character scalar.
}
\description{
Generate JavaScript for formatting the table
}
\keyword{internal}
