% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.get_unavailable_times_tz}
\alias{.get_unavailable_times_tz}
\title{Convert unavailable times to this user's timezone}
\usage{
.get_unavailable_times_tz(user_timezone)
}
\arguments{
\item{user_timezone}{The timezone currently chosen by the user (as a
character scalar).}
}
\value{
A tibble with columns day, hour, and unavailable.
}
\description{
Note: This currently depends on the global object `.unavailable_times()`.
}
\keyword{internal}
