% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.load_signups}
\alias{.load_signups}
\title{Load signup information from the google sheet}
\usage{
.load_signups(unavailable_times)
}
\arguments{
\item{unavailable_times}{A datetime vector of unavailble times, in UTC.}
}
\value{
A data.frame with information about user signups, in UTC.
}
\description{
Load signup information from the google sheet
}
\keyword{internal}
