% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{.load_unavailable_times}
\alias{.load_unavailable_times}
\title{Load Unavailable Times}
\usage{
.load_unavailable_times()
}
\value{
A one-column tibble of unavailable times.
}
\description{
Load Unavailable Times
}
\keyword{internal}
