% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttm_utils.R
\name{.make_single_utc}
\alias{.make_single_utc}
\title{Make a single UTC time}
\usage{
.make_single_utc(day, hour, timezone)
}
\arguments{
\item{day}{A single day.}

\item{hour}{A single hour.}

\item{timezone}{A single timezone.}
}
\value{
A single UTC time, equivalent to that day and time "next week."
}
\description{
Make a single UTC time
}
\keyword{internal}
