% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{.make_utc}
\alias{.make_utc}
\title{Make a UTC time vector}
\usage{
.make_utc(day, hour, timezone)
}
\arguments{
\item{day}{A vector of days.}

\item{hour}{A vector of hours.}

\item{timezone}{A vector of starting timezones.}
}
\value{
A vector of equivalent UTC times "next week."
}
\description{
Make a UTC time vector
}
\keyword{internal}
