% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{.next_weekday}
\alias{.next_weekday}
\title{Choose the next date of a given weekday name.}
\usage{
.next_weekday(target_weekday, tz)
}
\arguments{
\item{target_weekday}{The name of the weekday.}

\item{tz}{The timezone in which the day should be returned.}
}
\value{
A POSIXct of the next occurrance of that day.
}
\description{
Choose the next date of a given weekday name.
}
\keyword{internal}
