% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.preset_table}
\alias{.preset_table}
\title{Set up the table for this user}
\usage{
.preset_table(user_id, user_timezone, signups)
}
\arguments{
\item{user_id}{Slack's unique id for this user.}

\item{user_timezone}{The timezone currently chosen by the user (as a
character scalar).}

\item{signups}{A dataframe of signup information, converted to UTC.}
}
\value{
A data.frame for use in [rhandsontable::rhandsontable()].
}
\description{
Set up the table for this user
}
\keyword{internal}
