% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{.read_gs4}
\alias{.read_gs4}
\title{Read a Sheet from the GS4 Workbook}
\usage{
.read_gs4(...)
}
\arguments{
\item{...}{Arguments passed on to [googlesheets4::read_sheet()].}
}
\value{
A google sheet.
}
\description{
Read a Sheet from the GS4 Workbook
}
\keyword{internal}
