% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{.save_availability}
\alias{.save_availability}
\title{Save User Selections}
\usage{
.save_availability(user_name, user_id, book_name, timezone, time_table)
}
\arguments{
\item{user_name}{The name of this user.}

\item{user_id}{The Slack id of this user.}

\item{book_name}{The name of the book for which they're choosing times.}

\item{timezone}{The user's timezone.}

\item{time_table}{The user's selections.}
}
\value{
The sheet ID, invisibly.
}
\description{
Save User Selections
}
\keyword{internal}
