% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared-parameters}
\alias{.shared-parameters}
\title{Parameters used in multiple functions}
\arguments{
\item{id}{The identifier of this input.}

\item{session}{The shiny session. I don't know of a case where we shouldn't
use the default.}

\item{user_id}{Slack's unique id for this user.}

\item{user_name}{The Slack display name of this user.}

\item{user_timezone}{The timezone currently chosen by the user (as a
character scalar).}

\item{selected_book}{The book currently selected by this user (as a character
scalar).}

\item{signups}{A dataframe of signup information, converted to UTC.}
}
\description{
These are parameters that are likely to show up in multiple functions, so I
gather their definitions in one easy-to-find place.
}
\keyword{internal}
