% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{.timezone_server}
\alias{.timezone_server}
\title{A module for setting timezones.}
\usage{
.timezone_server(id = "timezone")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
A [shiny::moduleServer()] to update the timezone input using
  javascript.
}
\description{
A module for setting timezones.
}
\keyword{internal.}
