% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{.timezone_ui}
\alias{.timezone_ui}
\title{UI to Select a Timezone}
\usage{
.timezone_ui(id = "timezone")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
A [shiny::tagList()] that adds the appropriate info to the app for
  timezone settings.
}
\description{
UI to Select a Timezone
}
\keyword{internal}
