% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{.tz_minutes}
\alias{.tz_minutes}
\title{Find the Minutes Offset of a Timezone}
\usage{
.tz_minutes(timezone)
}
\arguments{
\item{timezone}{The timezone to check.}
}
\value{
The integer minutes difference between the supplied timezone and the
  same time in UTC.
}
\description{
Find the Minutes Offset of a Timezone
}
\keyword{internal}
