% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{.user_server}
\alias{.user_server}
\title{User Server Side}
\usage{
.user_server(id = "user_name")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
A moduleServer object.
}
\description{
User Server Side
}
\keyword{internal}
