% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(local = interactive(), ...)
}
\arguments{
\item{local}{Logical indicating whether to run the app locally or on a
server. Useful for debugging.}

\item{...}{arguments to pass to golem_opts. See
\code{\link[golem]{get_golem_options}} for more details.}
}
\description{
Run the Shiny Application
}
