% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookclubber-package.R
\docType{package}
\name{bookclubber-package}
\alias{bookclubber}
\alias{bookclubber-package}
\title{bookclubber: Schedule DSLC Book Clubs}
\description{
The Data Science Learning Community hosts a number of book clubs in which learners read data science books together, generally with one learner leading the discussion each week. This package provides a shiny app to help run these clubs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/bookclubber}
  \item Report bugs at \url{https://github.com/r4ds/bookclubber/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Priyanka Gagneja \email{priyankaigit@gmail.com}
}

}
\keyword{internal}
