% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{choose_time}
\alias{choose_time}
\title{Pick a Convenient Time}
\usage{
choose_time(book_name, facilitator_id, r4ds_tz = "America/Chicago")
}
\arguments{
\item{book_name}{The abbreviation for an approved book, such as rpkgs or
mshiny.}

\item{facilitator_id}{The Slack ID of the facilitator for this group.
Available via the user's "full profile" in Slack.}

\item{r4ds_tz}{A timezone in which to show the result, by default
"America/Chicago" since that's what the R4DS calendar is in.}
}
\value{
The best times, as a tibble.
}
\description{
There's a lot of work to do on this still.
}
