% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{choose_time}
\alias{choose_time}
\title{Pick a Convenient Time}
\usage{
choose_time(book_name)
}
\arguments{
\item{book_name}{The abbreviation for an approved book, such as rpkgs or
mshiny.}
}
\value{
The best time.
}
\description{
There's a lot of work to do on this still.
}
