% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.app_global}
\alias{.app_global}
\title{Stuff to run at startup}
\usage{
.app_global()
}
\description{
I'll have to see what should universally run at startup (eg, definitions of
the possible slots) vs what should be checked during use (eg, we should load
the "used" slots from time to time, although we can probably do that in
global with a reactivePoll, so... yeah, this will get more stuff).
}
\details{
After reading up a bit on the golem github, it looks like this isn't QUITE
what I want. We'll explore this further as we go, but likely we just want
most of this in server, with maybe a little bit in the golem_opts in run_app.
}
\keyword{internal}
