% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{.book_server}
\alias{.book_server}
\title{Book module server}
\usage{
.book_server(id = "book")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
The selected book as a reactive.
}
\description{
Book module server
}
\keyword{internal}
