% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_global.R
\name{.check_club_sheet}
\alias{.check_club_sheet}
\title{Check Club Sheet Modified Time}
\usage{
.check_club_sheet()
}
\value{
A string representing when the sheet was modified.
}
\description{
Check Club Sheet Modified Time
}
\keyword{internal}
