% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.format_table}
\alias{.format_table}
\title{Add a JavaScript formatter to the table}
\usage{
.format_table(hot, user_timezone, signups)
}
\arguments{
\item{hot}{A [rhandsontable::rhandsontable()] object.}

\item{user_timezone}{The timezone currently chosen by the user (as a
character scalar).}

\item{signups}{A dataframe of signup information, converted to UTC.}
}
\value{
A [rhandsontable::rhandsontable()] with JavaScript for formatting.
}
\description{
Add a JavaScript formatter to the table
}
\keyword{internal}
