% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{.save_availability}
\alias{.save_availability}
\title{Save User Selections}
\usage{
.save_availability(
  user_name,
  user_id,
  selected_book,
  user_timezone,
  time_table
)
}
\arguments{
\item{user_name}{The Slack display name of this user.}

\item{user_id}{Slack's unique id for this user.}

\item{selected_book}{The book currently selected by this user (as a character
scalar).}

\item{user_timezone}{The timezone currently chosen by the user (as a
character scalar).}

\item{time_table}{The user's selections.}
}
\value{
The sheet ID, invisibly.
}
\description{
Save User Selections
}
\keyword{internal}
