% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{.signups_to_matrix}
\alias{.signups_to_matrix}
\title{Summarize, pivot, and matricize signups}
\usage{
.signups_to_matrix(signups, user_timezone)
}
\arguments{
\item{signups}{The result of [.load_book_signups()].}

\item{user_timezone}{The timezone currently chosen by the user (as a
character scalar).}
}
\value{
A matrix of signups, with an integer count in each cell.
}
\description{
Summarize, pivot, and matricize signups
}
\keyword{internal}
