% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{.timezone_server}
\alias{.timezone_server}
\title{A module for setting timezones.}
\usage{
.timezone_server(id = "timezone", allowed_zones = OlsonNames())
}
\arguments{
\item{id}{The identifier of this input.}

\item{allowed_zones}{A character vector of allowed timezones.}
}
\value{
The selected timezone as a reactive.
}
\description{
A module for setting timezones.
}
\keyword{internal.}
