% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{.user_server}
\alias{.user_server}
\title{User module server}
\usage{
.user_server(id = "user_name")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
The user's `display_name` and `user_id` in a reactive list.
}
\description{
User module server
}
\keyword{internal}
