% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{.user_server}
\alias{.user_server}
\title{User Server Side}
\usage{
.user_server(id = "user_name", slack_user_info)
}
\arguments{
\item{id}{ID of this user displayer.}

\item{slack_user_info}{The object that contains info from Slack.}
}
\value{
A moduleServer object.
}
\description{
User Server Side
}
\keyword{internal}
