% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{.user_ui}
\alias{.user_ui}
\title{User UI Side}
\usage{
.user_ui(id = "user_name")
}
\arguments{
\item{id}{The identifier of this input.}
}
\value{
A shiny module.
}
\description{
User UI Side
}
\keyword{internal}
